third_party/blink/web_tests/transforms/3d/hit-testing/hover-rotated-with-children-negative-z-expected.txt

transformed
child 1
non transformed child
child 2

Checks that hit testing is correct when a transformed element has childs with negative z-offset

Element at 70, 70 has id "transformed": PASS
Element at 630, 130 has id "transformed": PASS
Element at 40, 130 has id "transformed": PASS
Element at 620, 270 has id "transformed": PASS
Element at 130, 100 has id "child1": PASS
Element at 200, 110 has id "child2": PASS
